requirements.txt
appnope==0.1.0
argcomplete==1.9.2
awscli==1.11.161
awsebcli==3.11.0
backports-abc==0.5
base58==0.2.4
bleach==2.1.1
blessed==1.14.2
boto3==1.4.7
botocore==1.7.19
cement==2.8.2
certifi==2017.7.27.1
cfn-flip==0.2.5
chardet==3.0.4
click==6.7
colorama==0.3.7
croniter==0.3.19
cycler==0.10.0
decorator==4.1.2
Django==1.11.4
django-debug-toolbar==1.8
django-filter==1.0.4
django-model-utils==3.0.0
django-redis==4.8.0
django-rq==0.9.6
django-rq-dashboard==0.3.3
django-storages==1.6.5
djangorestframework==3.6.4
dnspython==1.15.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
durationpy==0.5
email-validator==1.0.2
entrypoints==0.2.3
Faker==0.8.1
future==0.16.0
hjson==3.0.1
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.1.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pandocfilters==1.4.2
parso==0.1.0
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
placebo==0.8.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pyasn1==0.3.6
pyfcm==1.4.2
Pygments==2.2.0
pyparsing==2.2.0
pytesseract==0.1.7
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
redis==2.10.6
requests==2.9.1
requests-toolbelt==0.8.0
rq==0.8.2
rq-scheduler==0.7.0
rsa==3.4.2
s3transfer==0.1.11
scikit-learn==0.19.0
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
tabulate==0.7.5
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
toml==0.9.3
tornado==4.5.2
tqdm==4.19.1
traitlets==4.3.2
troposphere==2.1.0
typing==3.6.2
UkPostcodeParser==1.1.1
Unidecode==0.4.21
Unipath==1.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
